requirements.txt
streamlit==1.40.0
supabase==2.7.4
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1
